adjusted modules and setup proper default wireless floating winow behavior
I rearanged the the cpu ram and temp modules to the right because it was
too cluttered with the workspace module. I also  gave these three
modules some margin on the last child so that it can retain a sense of
seperation from the other modules.

Removed the tray module for now due to the blueman applet randomly
showing up and i would much rather use the waybar bluetooth module

Finally I didn't like the default behavior whenever a clicked module
spawned a window so i added some window rules to launch a floating
window in the middle of the screen